server/pyproject.toml
[tool.poetry]
name = "ProjectName"
version = "0.1.0"
description = ""
authors = ["colblind"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
django-filter = "^23.2"
django-environ = "^0.10.0"
psycopg = {extras = ["binary"], version = "^3.1.10"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
